exercise_8/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.11.10
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.20
redis==5.0.1
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.35.10
reportlab==4.2.0
markdown==3.6
asyncpg==0.30.0
SQLAlchemy==2.0.43
pgvector==0.4.1
psycopg2-binary==2.9.10